# ethiso-payment-service/src/main/resources/public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Payment Service API",
        "description": "These methods provide access to the Emerald payments API.",
        "version": "1.0.0"
    },
    "schemes": [
        "http"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/bic": {
            "patch": {
                "summary": "Register BIC for this party",
                "description": "Register a BIC to this party",
                "parameters": [
                    {
                        "name": "bic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RegisterBicBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true if the bic was registered",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                },
                "tags": [
                    "Sync"
                ]
            }
        },
        "/trustlines/{bic}/{ccy}/allow": {
            "patch": {
                "summary": "Set the allowed holdings for the given trustline",
                "description": "Set the allowed holdings between us and **party** for the given **ccy**. **allow** must be >= 0. The allowed holdings represents how much I am willing to owe you, this means I can prevent you from forcing me to hold your money (think of the Anti Money Laundering implications without this control).",
                "parameters": [
                    {
                        "name": "bic",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BIC"
                        }
                    },
                    {
                        "name": "ccy",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "allow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SetAllowBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Allowed holdings set",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "422": {
                        "description": "Set allowed holdings failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Sync"
                ]
            }
        },
        "/async/trustlines/{bic}/{ccy}/allow": {
            "patch": {
                "summary": "Set the allowed holdings for the given trustline asynchronously",
                "description": "Set the allowed holdings between us and **party** for the given **ccy** asynchronously. **allow** must be >= 0. The allowed holdings represents how much I am willing to owe you, this means I can prevent you from forcing me to hold your money (think of the Anti Money Laundering implications without this control).",
                "parameters": [
                    {
                        "name": "bic",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BIC"
                        }
                    },
                    {
                        "name": "ccy",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "allow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SetAllowBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true always",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "422": {
                        "description": "Set allowed holdings failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Async"
                ]
            }
        },
        "/trustlines/{bic}/{ccy}/limit": {
            "patch": {
                "summary": "Set the limit for the given trustline",
                "description": "Set the limit between us and **party** for the given **ccy**. **limit** must be >= 0. The limit represents how much I am will you to owe me. Consider this from a credit risk perspective, what is the maximum I am willing to lose if **party** goes bust.",
                "parameters": [
                    {
                        "name": "bic",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BIC"
                        }
                    },
                    {
                        "name": "ccy",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SetLimitBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Limit set",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "422": {
                        "description": "Set limit failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Sync"
                ]
            }
        },
        "/async/trustlines/{bic}/{ccy}/limit": {
            "patch": {
                "summary": "Set the limit for the given trustline asynchronously",
                "description": "Set the limit between us and **party** for the given **ccy** asynchronously. **limit** must be >= 0. The limit represents how much I am will you to owe me. Consider this from a credit risk perspective, what is the maximum I am willing to lose if **party** goes bust.",
                "parameters": [
                    {
                        "name": "bic",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BIC"
                        }
                    },
                    {
                        "name": "ccy",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SetLimitBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true always",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "422": {
                        "description": "Set limit failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Async"
                ]
            }
        },
        "/trustlines/{bic}/{ccy}/balance": {
            "get": {
                "summary": "Get the balance for the given trustline",
                "description": "Get the balance between us and **party** for the given **ccy**.",
                "parameters": [
                    {
                        "name": "bic",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BIC"
                        }
                    },
                    {
                        "name": "ccy",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The balance",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Get balance failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Sync"
                ]
            }
        },
        "/payments": {
            "patch": {
                "summary": "Send a payment",
                "description": "Send a payment from the given **from** IBAN (must be registered to us) to the **to** IBAN (must be registered to a party and that party can't be us) for the **amount** in currency **ccy**. The **remittanceInfo** is sent across the network along with the rest of the payment info and is transmitted to the beneficiary.",
                "parameters": [
                    {
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PaymentBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment sent",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "422": {
                        "description": "Payment failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Sync"
                ]
            }
        },
        "/async/payments": {
            "patch": {
                "summary": "Send a payment asynchronously",
                "description": "Send a payment from the given **from** IBAN (must be registered to us) to the **to** IBAN (must be registered to a party and that party can't be us) for the **amount** in currency **ccy** asynchronously. The **remittanceInfo** is sent across the network along with the rest of the payment info and is transmitted to the beneficiary.",
                "parameters": [
                    {
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PaymentBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true always",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "422": {
                        "description": "Payment failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "tags": [
                    "Async"
                ]
            }
        }
    },
    "definitions": {
        "CCY": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "example": "EUR"
        },
        "Amount": {
            "type": "string",
            "pattern": "^[1-9][0-9]*/.[0-9]{2}$",
            "example": "100.00"
        },
        "IBAN": {
            "type": "string",
            "pattern": "^[A-Z]{2}[0-9]{2}[A-Z0-9]{11,30}$",
            "example": "GB29NWBK60161331926819"
        },
        "BIC": {
            "type": "string",
            "pattern": "^[A-Z0-9]{4}([A-Z]{2}[A-Z0-9]{2}([A-Z0-9]{3})?)?$",
            "example": "RBOSGB2L"
        },
        "RegisterBicBody": {
            "type": "object",
            "properties": {
                "bic": {
                    "$ref": "#/definitions/BIC"
                }
            },
            "required": [
                "bic"
            ]
        },
        "SetAllowBody": {
            "type": "object",
            "properties": {
                "allow": {
                    "$ref": "#/definitions/Amount"
                }
            },
            "required": [
                "allow"
            ]
        },
        "SetLimitBody": {
            "type": "object",
            "properties": {
                "limit": {
                    "$ref": "#/definitions/Amount"
                }
            },
            "required": [
                "limit"
            ]
        },
        "PaymentBody": {
            "type": "object",
            "properties": {
                "fromBic": {
                    "$ref": "#/definitions/BIC"
                },
                "toBic": {
                    "$ref": "#/definitions/BIC"
                },
                "fromIban": {
                    "$ref": "#/definitions/IBAN"
                },
                "toIban": {
                    "$ref": "#/definitions/IBAN"
                },
                "ccy": {
                    "$ref": "#/definitions/CCY"
                },
                "amount": {
                    "$ref": "#/definitions/Amount"
                },
                "remittanceInfo": {
                    "type": "string",
                    "example": "MyRef123"
                }
            },
            "required": [
                "fromBic",
                "toBic",
                "fromIban",
                "toIban",
                "ccy",
                "amount",
                "remittanceInfo"
            ]
        },
        "ConfigResponse": {
            "type": "object",
            "properties": {
                "enode": {
                    "type": "string"
                },
                "registrarContractAddress": {
                    "type": "string"
                },
                "ready": {
                    "type": "boolean"
                }
            },
            "required": [
                "enode",
                "registrarContractAddress",
                "ready"
            ]
        }
    }
}